+++ src/index.js
@@ -53,7 +53,7 @@ export default function fetch (url, opts = {}) {
     if (request.useElectronNet) {
       headers = options.headers
       delete options.headers
-      options.session = opts.session || electron.session.fromPartition('electron-fetch')
+      options.session = opts.session || electron.session.defaultSession // we have to use a persistent session here, because of https://github.com/electron/electron/issues/13587
     }
     const req = send(options)
     if (request.useElectronNet) {